requirements.txt
diffusers==0.34.0
numpy==2.2.6
Pillow==11.2.1
torch==2.6.0
torchvision==0.21.0
transformers==4.53.0
